Add asteriod and medkit models

+++ Assets/Scripts/GameEvents.cs
@@ -5,9 +5,9 @@ public class GameEvents : MonoBehaviour {
 
 	public static GameEvents current;
 
-	public Action<int> HP_CHANGED = i => { };
+	public Action<float> HP_CHANGED = i => { };
 	public Action<int> SCORE_CHANGED = i => { };
-	public Action<int> HP_CHANGE = i => { };
+	public Action<float> HP_CHANGE = i => { };
 	public Action<int> SCORE_CHANGE = i => { };
 	public Action 	   PLAYER_DEAD = () => { };
